WIP: Refactor MPDClient
This should make handling the queuing side work more reliably.
