+++ Persephone.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,9 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:../README.md">
-   </FileRef>
-   <FileRef
       location = "self:">
    </FileRef>
 </Workspace>
