+++ Persephone/Preferences/Controllers/AlbumArtPrefsController.swift
@@ -9,9 +9,29 @@
 import Cocoa
 
 class AlbumArtPrefsController: NSViewController {
+  var preferences = Preferences()
+
   override func viewDidLoad() {
     super.viewDidLoad()
 
+    if let mpdLibraryDir = preferences.mpdLibraryDir {
+      mpdLibraryDirField.stringValue = mpdLibraryDir
+    }
+
     preferredContentSize = NSMakeSize(view.frame.size.width, view.frame.size.height)
   }
+
+  override func viewDidAppear() {
+    super.viewDidAppear()
+
+    guard let title = title
+      else { return }
+    self.parent?.view.window?.title = title
+  }
+
+  @IBAction func updateMpdLibraryDir(_ sender: NSTextField) {
+    preferences.mpdLibraryDir = sender.stringValue
+  }
+
+  @IBOutlet var mpdLibraryDirField: NSTextField!
 }
